pyproject.toml
[tool.poetry]
name = "polyhistor"
version = "0.1.0"
description = ""
authors = ["jeff3071 <jeff563715@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
langchain = "^0.1.13"
pymupdf = "^1.24.0"
tqdm = "^4.66.2"
markdown = "^3.6"
rank-bm25 = "^0.2.2"
pre-commit = "^3.7.0"
qdrant-client = "^1.9.0"
python-docx = "^1.1.2"
python-dotenv = "^1.0.1"
unstructured = "^0.13.7"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
torch = {url = "https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.2.2%2Bcpu.cxx11.abi-cp311-cp311-linux_x86_64.whl"}
sentence-transformers = "^3.0.1"


[tool.poetry.group.test.dependencies]
vcrpy = "^6.0.1"
pytest = "^8.2.2"
syrupy = "^4.6.1"
pytest-asyncio = "^0.23.7"
pytest-recording = "^0.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
